Move requests code to separate file
